【问题】 路径错误

+++ shop-order/src/main/java/com/chow/dao/OrderDao.java
@@ -1,8 +1,7 @@
 package chow.dao;
 
 import com.chow.domain.Order;
-import com.chow.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderDao extends JpaRepository<Order, Integer> {
+public interface OrderDao extends JpaRepository<Order, Long> {
 }
